+++ changelog.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 23rd August 2024
+
+- Enhancement of the `regExpTemplate` function to accommodate # comments.
+
 ## 15th May 2024
 
 - New simple `range` function added to the ranges module.